GROW-979 Type is defined as int in the DB

+++ modules/social-activity-customizer/social-activity-customizer-activities-query-helper/src/main/java/com/liferay/social/activity/customizer/query/helper/CustomSocialActivitiesQueryHelper.java
@@ -72,7 +72,7 @@ public class CustomSocialActivitiesQueryHelper
 		_classicPaginationStart = classicPaginationStart;
 	}
 
-	public void setTypes(long[] types) {
+	public void setTypes(int[] types) {
 		_types = types;
 	}
 
@@ -101,6 +101,6 @@ public class CustomSocialActivitiesQueryHelper
 	private int _classicPaginationStart;
 	private CustomSocialActivitySetLocalService
 		_customSocialActivitySetLocalService;
-	private long[] _types;
+	private int[] _types;
 
 }
